Some Cleanup 12/02/2022
Comments were added for the Network's cost().
MNIST values converted from 0-255 to 0-1.
Removed Layer's calculate()'s conversion of all inputs to decimals.

+++ src/MnistDatabase.java
@@ -102,12 +102,12 @@ public class MnistDatabase {
      * @param imageNum
      * @return
      */
-    public int[][] getTrainingImageMatrix(int imageNum) {
+    public double[][] getTrainingImageMatrix(int imageNum) {
 
         assert imageNum < training.length: " imageNum out of bounds.";
 
         MnistMatrix matrix = training[imageNum];
-        int[][] pixels = new int[matrix.getNumberOfRows()][matrix.getNumberOfColumns()];
+        double[][] pixels = new double[matrix.getNumberOfRows()][matrix.getNumberOfColumns()];
 
         for (int r = 0; r < matrix.getNumberOfRows(); r++ ) {
             for (int c = 0; c < matrix.getNumberOfColumns(); c++) {
@@ -170,12 +170,12 @@ public class MnistDatabase {
      * @param imageNum
      * @return
      */
-    public int[][] getTestingImageMatrix(int imageNum) {
+    public double[][] getTestingImageMatrix(int imageNum) {
 
         assert imageNum < testing.length: " imageNum out of bounds.";
 
         MnistMatrix matrix = testing[imageNum];
-        int[][] pixels = new int[matrix.getNumberOfRows()][matrix.getNumberOfColumns()];
+        double[][] pixels = new double[matrix.getNumberOfRows()][matrix.getNumberOfColumns()];
 
         for (int r = 0; r < matrix.getNumberOfRows(); r++ ) {
             for (int c = 0; c < matrix.getNumberOfColumns(); c++) {
